+++ src/utils/merge_deeply.ts
@@ -1,0 +1,44 @@
+// eslint-disable-next-line
+function mergeDeeply(target: any, source: any): any {
+  // eslint-disable-next-line
+  const isObject = (obj: any) =>
+    obj && typeof obj === "object" && !Array.isArray(obj);
+  const result = Object.assign({}, target);
+  if (Array.isArray(target) && isObject(source)) {
+    const resultArray = JSON.parse(JSON.stringify(target));
+    for (const [sourceKey, sourceValue] of Object.entries(source)) {
+      // @ts-ignore
+      const targetValue = target[sourceKey];
+
+      if (
+        isObject(sourceValue) &&
+        Object.prototype.hasOwnProperty.call(target, sourceKey)
+      ) {
+        if (isObject(targetValue)) {
+          resultArray[sourceKey] = mergeDeeply(targetValue, sourceValue);
+        } else {
+          resultArray[sourceKey] = sourceValue;
+        }
+      } else {
+        resultArray[sourceKey] = sourceValue;
+      }
+    }
+    return resultArray;
+  }
+  if (isObject(target) && isObject(source)) {
+    for (const [sourceKey, sourceValue] of Object.entries(source)) {
+      const targetValue = target[sourceKey];
+      if (
+        isObject(sourceValue) &&
+        Object.prototype.hasOwnProperty.call(target, sourceKey)
+      ) {
+        result[sourceKey] = mergeDeeply(targetValue, sourceValue);
+      } else {
+        Object.assign(result, { [sourceKey]: sourceValue });
+      }
+    }
+  }
+  return result;
+}
+
+export default mergeDeeply;
